Refactor - split into 3 projects
