+++ .gitignore
@@ -30,5 +30,3 @@ _ReSharper*/
 .vs/
 #Nuget packages folder
 packages/
-# generated by NSwag
-ManageCoursesApiClient.cs
