+++ src/api-client/manage-courses-api-swagger.json
