+++ src/importer/UcasCourseImporter.csproj
@@ -6,17 +6,16 @@
     <RootNamespace>GovUk.Education.ManageCourses.UcasCourseImporter</RootNamespace>
   </PropertyGroup>
 
-  <Target Name="NSwag" BeforeTargets="Compile">
-    <Copy SourceFiles="@(Reference)" DestinationFolder="$(OutDir)References" />
-    <Exec Command="$(NSwagExe_Core20) swagger2csclient /input:manage-courses-api-swagger.json /classname:ManageCoursesApiClient /namespace:GovUk.Education.ManageCourses.UcasCourseImporter /output:ManageCoursesApiClient.cs" />
-    <RemoveDir Directories="$(OutDir)References" />
-  </Target>
-
   <ItemGroup>
     <PackageReference Include="DotNetCore.NPOI" Version="1.0.2" />
     <PackageReference Include="Microsoft.Extensions.CommandLineUtils" Version="1.1.1" />
     <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
     <PackageReference Include="NSwag.MSBuild" Version="11.17.13" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\api-client\ApiClient.csproj" />
+    <ProjectReference Include="..\xls-reader\XlsReader.csproj" />
   </ItemGroup>
 
 </Project>
