+++ src/xls-reader/XlsReader.cs
@@ -1,41 +1,20 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using System.IO;
 using System.Linq;
-using Microsoft.Extensions.CommandLineUtils;
+using GovUk.Education.ManageCourses.ApiClient;
 using NPOI.HSSF.UserModel;
 
-namespace GovUk.Education.ManageCourses.UcasCourseImporter
+namespace GovUk.Education.ManageCourses.Xls
 {
-    class Program
+    public class XlsReader
     {
-        static void Main(string[] args)
+        public  List<Course> ReadCourses(string folder)
         {
-            var app = new CommandLineApplication();
-            var folderOption = app.Option("-$|-f|--folder <folder>", "Folder to read UCAS .xls files from ", CommandOptionType.SingleValue);
-            app.HelpOption("-?|-h|--help");
-            app.Execute(args);
-
-            var courses = ReadCourses(folderOption.Value());
-            SendToManageCoursesApi(courses);
-        }
-
-        private static void SendToManageCoursesApi(List<Course> courses)
-        {
-            Console.WriteLine("Posting to api...");
-            var client = new ManageCoursesApiClient();
-            var payload = new Payload { Courses = new ObservableCollection<Course>(courses) };
-            client.ImportAsync(payload).Wait();
-            Console.WriteLine("Done.");
-        }
-
-        private static List<Course> ReadCourses(string folder)
-        {
-            Console.Write("Reading course xls file from: ");
-            Console.WriteLine(folder);
-            var courses = new List<Course>();
             var file = new FileInfo(Path.Combine(folder, "GTTR_CRSE.xls"));
+            Console.WriteLine("Reading course xls file from: " + file.FullName);
+
+            var courses = new List<Course>();
             using (var stream = new FileStream(file.FullName, FileMode.Open))
             {
                 var wb = new HSSFWorkbook(stream);
